Migrated vehicle and cat subclasses to their own files
