+++ NPCs/Bosses/Godcat/Godcat_Creator.cs
@@ -1,0 +1,257 @@
+﻿using EBF.EbfUtils;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using System;
+using Terraria;
+using Terraria.Audio;
+using Terraria.GameContent.Bestiary;
+using Terraria.ID;
+using Terraria.ModLoader;
+
+namespace EBF.NPCs.Bosses.Godcat
+{
+    [AutoloadBossHead]
+    public class Godcat_Creator : Godcat_Vehicle
+    {
+        private Vector2 BarrelPos => NPC.Center + new Vector2(80 * NPC.direction, -16);
+        public override void SetStaticDefaults()
+        {
+            base.SetStaticDefaults();
+            Main.npcFrameCount[Type] = 1;
+
+            //Bestiary
+            NPCID.Sets.BossBestiaryPriority.Add(Type); //Grouped with other bosses
+            var drawModifiers = new NPCID.Sets.NPCBestiaryDrawModifiers()
+            {
+                CustomTexturePath = "EBF/Assets/Textures/Bestiary/Godcat_Preview",
+                PortraitScale = 0.6f, // Portrait refers to the full picture when clicking on the icon in the bestiary
+                PortraitPositionYOverride = 0f,
+            };
+            NPCID.Sets.NPCBestiaryDrawOffset.Add(Type, drawModifiers);
+        }
+        public override void SetDefaults()
+        {
+            base.SetDefaults();
+            NPC.width = 202;
+            NPC.height = 154;
+            NPC.HitSound = SoundID.NPCHit4;
+            NPC.DeathSound = SoundID.Item14;
+            idleTexture = ModContent.Request<Texture2D>(Texture);
+            currentTexture = idleTexture.Value;
+
+            stateDurations = new()
+            {
+                [State.Idle] = 200,
+                [State.TurningBallCircle] = 240,
+                [State.TurningBallSpiral] = 300,
+                [State.CreatorThunderBall] = 200,
+                [State.CreatorHolyDeathray] = 200,
+            };
+        }
+        public override void SetBestiary(BestiaryDatabase database, BestiaryEntry bestiaryEntry)
+        {
+            bestiaryEntry.Info.AddRange([
+                new MoonLordPortraitBackgroundProviderBestiaryInfoElement(), // Plain black background
+				new FlavorTextBestiaryInfoElement("Mods.EBF.Bestiary.Godcat_Creator")
+            ]);
+        }
+        public override void AI()
+        {
+            base.AI();
+            var player = Main.player[NPC.target];
+            switch (currentState)
+            {
+                case State.TurningBallCircle:
+                    CreateTurningBallsCircles();
+                    break;
+                case State.TurningBallSpiral:
+                    CreateTurningBallSpiral();
+                    break;
+                case State.CreatorThunderBall:
+                    CreateThunderBalls();
+                    break;
+                case State.CreatorHolyDeathray:
+                    CreateHolyDeathray();
+                    break;
+            }
+        }
+        public override void OnKill()
+        {
+            base.OnKill();
+            NPC.CreateExplosionEffect(EBFUtils.ExplosionSize.Large);
+        }
+        protected override void Move(Player player)
+        {
+            var offset = new Vector2(550, -100);
+            if (currentState == State.CreatorHolyDeathray)
+            {
+                offset = new Vector2(400, 16);
+            }
+
+            var preferredPosition = player.Center + offset;
+            NPC.Center = Vector2.Lerp(NPC.Center, preferredPosition, 0.03f);
+        }
+        protected override void BeginNextPhase(Player player)
+        {
+            var type = ModContent.NPCType<Godcat_Dark>();
+            var pos = player.Center.ToPoint() + new Point(NPC.direction * 1600, 0);
+            NPC.NewNPC(NPC.GetSource_FromAI(), pos.X, pos.Y, type, 0, Phase);
+        }
+        private void CreateTurningBallsCircles()
+        {
+            if (Main.GameUpdateCount % 30 == 0)
+            {
+                var amount = 12;
+                var speed = 5;
+                var type = ModContent.ProjectileType<Godcat_TurningBall>();
+                for (float theta = 0; theta < MathF.Tau; theta += MathF.Tau / amount)
+                {
+                    var velocity = Vector2.UnitX.RotatedBy(theta) * speed;
+                    Projectile.NewProjectile(NPC.GetSource_FromAI(), NPC.Center, velocity, type, NPC.damage, 3f, -1, (float)GodcatBallTypes.LightBig, -0.005f);
+
+                    if (IsAlone)
+                    {
+                        Projectile.NewProjectile(NPC.GetSource_FromAI(), NPC.Center, velocity * 0.9f, type, NPC.damage, 3f, -1, (float)GodcatBallTypes.LightBig, 0.005f);
+                    }
+                }
+
+                SoundEngine.PlaySound(SoundID.Item39, NPC.position); //Razorpine sound
+            }
+        }
+        private void CreateTurningBallSpiral()
+        {
+            if (Main.GameUpdateCount % 2 == 0)
+            {
+                var speed = 4;
+                var velocity = (Main.GameUpdateCount * 0.2f).ToRotationVector2() * speed;
+                var type = ModContent.ProjectileType<Godcat_TurningBall>();
+                Projectile.NewProjectile(NPC.GetSource_FromAI(), NPC.Center, velocity, type, NPC.damage, 3f, -1, (float)GodcatBallTypes.LightBig, -0.005f);
+
+                SoundEngine.PlaySound(SoundID.Item39, NPC.position); //Razorpine sound
+            }
+
+            if (IsAlone && Main.GameUpdateCount % 60 == 0)
+            {
+                var amount = 12;
+                var speed = 6;
+                var type = ModContent.ProjectileType<Godcat_LightDiamond>();
+                for (float theta = 0; theta < MathF.Tau; theta += MathF.Tau / amount)
+                {
+                    var velocity = Vector2.UnitX.RotatedBy(theta) * speed;
+                    Projectile.NewProjectile(NPC.GetSource_FromAI(), NPC.Center, velocity, type, NPC.damage, 3f);
+                }
+
+                SoundEngine.PlaySound(SoundID.Item72, NPC.position); //Shadowbeam sound
+            }
+        }
+        private void CreateThunderBalls()
+        {
+            //Three small bursts
+            if (StateTimer == 0 || StateTimer == 66 || StateTimer == 133)
+            {
+                SoundEngine.PlaySound(SoundID.Item8, NPC.position);
+
+                //Form a ring of thunder balls
+                var amount = 24;
+                if (!IsAlone)
+                {
+                    amount = 18;
+                }
+
+                var delay = 40;
+                var type = ModContent.ProjectileType<Creator_Thunderball>();
+                for (float theta = 0; theta < MathF.Tau; theta += MathF.Tau / amount)
+                {
+                    var velocity = theta.ToRotationVector2();
+                    velocity.Y *= 0.4f;
+
+                    //Some balls should draw behind creator, we send that info via ai[1]
+                    var drawBehind = 0;
+                    if (theta > MathF.PI) // ">" because theta goes clockwise for some reason?
+                    {
+                        drawBehind = 1;
+                    }
+
+                    //ai[0] is how long it takes before the balls launch
+                    //ai[2] is the owner, real owner must be -1 for dmg to work, owner is used to keep balls attached until they launch
+                    Projectile.NewProjectile(NPC.GetSource_FromAI(), NPC.Center, velocity, type, NPC.damage, 3f, -1, delay, drawBehind, NPC.whoAmI);
+                    delay += 2;
+                }
+            }
+
+            //Final huge burst
+            else if (StateTimer == 199)
+            {
+                SoundEngine.PlaySound(SoundID.Item8, NPC.position);
+
+                //Form a ring of thunder balls
+                var amount = 8;
+                var delay = 40;
+                var type = ModContent.ProjectileType<Creator_HugeThunderball>();
+                for (float theta = 0; theta < MathF.Tau; theta += MathF.Tau / amount)
+                {
+                    var velocity = theta.ToRotationVector2();
+                    velocity.Y *= 0.4f;
+
+                    //Some balls should draw behind creator, we send that info via ai[1]
+                    var drawBehind = 0;
+                    if (theta > MathF.PI) // ">" because theta goes clockwise for some reason?
+                    {
+                        drawBehind = 1;
+                    }
+
+                    //ai[0] is how long it takes before the balls launch
+                    //ai[2] is the owner, real owner must be -1 for dmg to work, owner is used to keep balls attached until they launch
+                    Projectile.NewProjectile(NPC.GetSource_FromAI(), NPC.Center, velocity, type, NPC.damage, 3f, -1, delay, drawBehind, NPC.whoAmI);
+                    delay += 10;
+                }
+            }
+        }
+        private void CreateHolyDeathray()
+        {
+            if (StateTimer == 0)
+            {
+                SoundEngine.PlaySound(SoundID.NPCDeath58, NPC.Center);
+            }
+
+            if (StateTimer < 150)
+            {
+                ChargeUpDust();
+            }
+            else if (StateTimer == 150)
+            {
+                ShootLaser();
+
+                //Create arc of diamonds
+                var speed = 4f;
+                var amount = 4;
+                var spread = 1f;
+                var type = ModContent.ProjectileType<Godcat_LightDiamond>();
+                var dir = (NPC.direction == 1 ? 0 : MathHelper.Pi).ToRotationVector2();
+                for (float theta = -spread; theta < spread; theta += 2 * spread / amount)
+                {
+                    var velocity = dir.RotatedBy(theta) * Main.rand.NextFloat(0.9f, 1.1f) * speed;
+                    Projectile.NewProjectile(NPC.GetSource_FromAI(), NPC.Center, velocity, type, NPC.damage, 3f);
+                }
+            }
+        }
+        private void ShootLaser()
+        {
+            var velocity = new Vector2(NPC.direction, 0);
+            var type = ModContent.ProjectileType<Creator_HolyDeathray>();
+            Projectile.NewProjectile(NPC.GetSource_FromAI(), NPC.Center, velocity, type, NPC.damage, 3, -1, NPC.whoAmI);
+
+            var sound = SoundID.Zombie104; // Moon Lord deathray sound
+            sound.Pitch = 1.4f;
+            sound.Volume = 0.3f;
+            SoundEngine.PlaySound(sound, NPC.Center);
+        }
+        private void ChargeUpDust()
+        {
+            var pos = BarrelPos + new Vector2(NPC.direction, 0).RotatedByRandom(0.5f) * 32;
+            var vel = pos.DirectionTo(BarrelPos);
+            var dust = Dust.NewDustPerfect(pos, DustID.AncientLight, vel, 0, default, 2.0f);
+            dust.noGravity = true;
+        }
+    }
+}
